// src/lib/reward-api/models/data_CreateFinanceDocRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { data_ApplicationDetailItemVO } from './data_ApplicationDetailItemVO';
export type data_CreateFinanceDocRequest = {
    application_detail: Array<data_ApplicationDetailItemVO>;
    creator?: string;
    description?: string;
    project_id: number;
};